+++ crates/rust-releases-rust-dist/src/client.rs
@@ -1,0 +1,13 @@
+use rust_releases_io::{BoxFuture, Document};
+
+pub trait DistIndexClient {
+    type Error;
+
+    fn download(&self) -> Result<Document, Self::Error>;
+}
+
+pub trait AsyncDistIndexClient: Send + Sync {
+    type Error;
+
+    fn download(&self) -> BoxFuture<'_, Result<Document, Self::Error>>;
+}